+++ CHANGES.md
@@ -1,6 +1,7 @@
 master (in development)
 =======================
 
+- Add `fancy-battery-show-percentage`
 - Add `fancy-battery-status-update-functions`
 - Fix detection of charging battery